requirements.txt
certifi==2020.12.5
chardet==3.0.4
contextvars==2.4
google-trans-new==1.1.9
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.15
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.2
pkg-resources==0.0.0
pytesseract==0.3.7
requests==2.25.1
rfc3986==1.4.0
sniffio==1.2.0
urllib3==1.26.4
